Push to a branch instead of a fork.
Many of our repos run integration / system tests on Travis. These
require access to credentials to run, so they must be run from a branch
instead of fork.

Also, so that we aren't creating a new branch every day, push to a
constant branch name instead of a new branch. (I forget why I wasn't
doing this in the first place.)